+++ src/main/java/org/mariadb/jdbc/message/client/PrepareExecutePacket.java
@@ -103,7 +103,7 @@ public final class PrepareExecutePacket implements RedoableWithPrepareClientMess
         writer.writeByte(p.getBinaryEncodeType());
         writer.writeByte(0);
         if (p.isNull()) {
-          nullBitsBuffer[i / 8] |= (1 << (i % 8));
+          nullBitsBuffer[i / 8] |= (byte) (1 << (i % 8));
         }
       }
